fix(billing): rename meter helper off seed* prefix for wiring gate
validate:seed-wiring requires every scripts/setup/seed*.ts to have an
npm script. stripe-billing-meter is a library module (like
stripe-price-match), not a seed CLI.
